Add "team" to totalfilter in dstat_net_packets.py
In RHEL7, Network Teaming has been chosen to refer to a new
implementation of "bonding", and as far as I know, we prefer using
teaming instead of bonding in our new environment.

When using dstat to monitor network streaming, we found that dstat can't
handle "team" card. For example, we use eth0 & eth1 for teaming, running
net tests, the results which dstat return would be 118MB/s in --net/eht0--,
112MB/s in --net/team1--, and 230MB/s in --net/total--.

Signed-off-by: Li Hongjie lihongjie@cmss.chinamobile.com